# thinx-api-openapi.yaml
openapi: 3.0.3
info:
  title: THiNX Device Management API
  description: |
    THiNX Device Management API provides functionality for managing IoT devices,
    firmware updates, and related services. The API supports both v1 and v2 endpoints.
  version: 2.0.0
  contact:
    name: THiNX Support
    url: https://thinx.cloud
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://rtm.thinx.cloud/api
    description: Legacy API (v1)
  - url: https://rtm.thinx.cloud/api/v2
    description: Current API (v2)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    sessionCookie:
      type: apiKey
      in: cookie
      name: x-thx-core
      description: Session cookie obtained after login
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authentication
      description: API key for device authentication
    
    ownerApiKeyAuth:
      type: object
      description: Owner ID and API key in request body
      properties:
        owner:
          type: string
          description: Owner ID
        api_key:
          type: string
          description: API key
...
paths:
...
  /mesh:
    get:
      summary: List meshes
      description: List all mesh networks for the authenticated user
      tags:
        - Mesh
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        '200':
          description: List of mesh networks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  meshes:
                    type: array
                    items:
                      type: object
                      properties:
                        mesh_id:
                          type: string
                        alias:
                          type: string
        '401':
          description: Unauthorized

    put:
      summary: Create mesh
      description: Create a new mesh network
      tags:
        - Mesh
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - alias
              properties:
                alias:
                  type: string
      responses:
        '200':
          description: Mesh created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  mesh_id:
                    type: string
        '401':
          description: Unauthorized

    delete:
      summary: Delete mesh
      description: Delete a mesh network
      tags:
        - Mesh
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - mesh_id
              properties:
                mesh_id:
                  type: string
      responses:
        '200':
          description: Mesh deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized

  # RSA Key endpoints
...
  /oauth/google/callback:
    get:
      summary: Google OAuth callback
      description: Handle OAuth callback from Google
      tags:
        - OAuth
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect after OAuth flow
...